[Update] Menubar fully functional #21

+++ src/app/integration/interpreter/interpreter.hpp
@@ -17,7 +17,7 @@ std::pair<bool, uint64_t> getRegister(std::string name);
 extern bool ucInit();
 extern uc_engine *uc;
 extern int regNameToConstant(std::string name);
-extern bool runCode(const std::string& code_in, int instructionCount);
+extern bool runCode(const std::string& code_in, uint64_t instructionCount);
 extern void showRegs();
 extern uintptr_t ENTRY_POINT_ADDRESS;
 extern uintptr_t MEMORY_ALLOCATION_SIZE;
